+++ ctrl/worker.go
@@ -22,11 +22,13 @@ import (
 	"fmt"
 	"net/url"
 	"path"
+	"strconv"
 	"strings"
 
 	"github.com/ezbastion/ezb_srv/cache"
-	"github.com/ezbastion/ezb_srv/model"
+	"github.com/ezbastion/ezb_srv/models"
 	"github.com/ezbastion/ezb_srv/tool"
+	"github.com/gin-contrib/location"
 
 	"github.com/gin-gonic/gin"
 	"github.com/go-resty/resty"
@@ -34,41 +36,47 @@ import (
 )
 
 func SendAction(c *gin.Context, storage cache.Storage) {
-	ac, _ := c.Get("action")
-	action := ac.(model.EzbActions)
-	p, _ := c.Get("params")
-	params := p.(map[string]string)
-	body := params["body"]
-	wk, _ := c.Get("worker")
-	worker := wk.(model.EzbWorkers)
-	cf, _ := c.Get("configuration")
-	conf := cf.(*model.Configuration)
-	rawpath := c.Request.URL.EscapedPath()
-	rawquery := c.Request.URL.RawQuery
-	tr, _ := c.Get("trace")
-	trace := tr.(model.EzbLogs)
-	exPath := c.MustGet("exPath").(string)
-	key := fmt.Sprintf("%x", md5.Sum([]byte(rawpath+rawquery+body)))
+	trace, ok := c.MustGet("trace").(models.EzbLogs)
+	if !ok {
+		err := "cant read context trace"
+		log.WithFields(log.Fields{"controller": "worker"}).Error(err)
+		c.JSON(500, err)
+		return
+	}
+
 	logg := log.WithFields(log.Fields{
 		"controller": "worker",
 		"xtrack":     trace.Xtrack,
 	})
+	logg.Debug("start")
+	action := c.MustGet("action").(models.EzbActions)
+	exPath := c.MustGet("exPath").(string)
+	tokenid := c.MustGet("tokenid").(string)
+	job := c.MustGet("job").(models.EzbJobs)
+	data := c.MustGet("params").(map[string]string)
+	worker := c.MustGet("worker").(models.EzbWorkers)
+	conf := c.MustGet("configuration").(*models.Configuration)
+	body := data["body"]
+	rawpath := c.Request.URL.EscapedPath()
+	rawquery := c.Request.URL.RawQuery
+	key := fmt.Sprintf("%x", md5.Sum([]byte(rawpath+rawquery+body)))
+	var meta models.EzbParamMeta
+	meta.Job = job
+	var params models.EzbParams
+	params.Data = data
+	params.Meta = meta
+
 	// worker.Request++
 	tool.IncRequest(&worker, c)
-	// var respStruct map[string]interface{}
 	var respStruct interface{}
 	if action.Jobs.Cache > 0 && c.Request.Method == "GET" {
-		response, ok := model.GetResult(storage, key)
+		response, ok := models.GetResult(storage, key)
 		if ok {
-			// log.Info("found")
 			js := strings.NewReader(string(response))
 			json.NewDecoder(js).Decode(&respStruct)
-			// respStruct["xtrack"] = trace.Xtrack
 			c.JSON(200, respStruct)
 			return
-			// break
 		}
-		// log.Info("not found")
 	}
 	var Url *url.URL
 	Url, err := url.Parse(worker.Fqdn)
@@ -77,39 +85,48 @@ func SendAction(c *gin.Context, storage cache.Storage) {
 		logg.Error(err)
 		c.JSON(500, err)
 		return
-		// break
 	}
 	cert, err := tls.LoadX509KeyPair(path.Join(exPath, conf.PublicCert), path.Join(exPath, conf.PrivateKey))
 	if err != nil {
 		logg.Error(err)
 		c.JSON(500, err.Error())
 		return
-		// break
 	}
 	resty.SetRootCertificate(path.Join(exPath, conf.CaCert))
 	resty.SetCertificates(cert)
 	resp, err := resty.R().
 		SetHeader("Accept", "application/json").
 		SetHeader("X-Track", trace.Xtrack).
+		SetHeader("X-Polling", strconv.FormatBool(action.Polling)).
+		SetHeader("x-ezb-tokenid", tokenid).
 		SetBody(&params).
 		SetResult(&respStruct).
 		Post(Url.String())
-	// trace.Status = resp.Status()
 	if err != nil {
-		logg.Error(err)
-		c.JSON(500, err)
+		logg.Warning(err)
+		c.JSON(resp.StatusCode(), err)
 		return
-		// break
 	}
 	if resp.StatusCode() < 300 {
-		c.JSON(resp.StatusCode(), respStruct)
+		var task models.EzbTasks
+		err = json.Unmarshal(resp.Body(), &task)
+		if action.Polling == true && err == nil {
+			u := location.Get(c)
+			logg.Debug("respStruct is models.EzbTasks")
+			Location := fmt.Sprintf("%s://%s/tasks/%04d%s/status", u.Scheme, u.Host, worker.ID, task.UUID)
+			c.Writer.Header().Set("Location", Location)
+			task.StatusURL = Location
+			task.LogURL = fmt.Sprintf("%s://%s/tasks/%04d%s/log", u.Scheme, u.Host, worker.ID, task.UUID)
+			task.ResultURL = fmt.Sprintf("%s://%s/tasks/%04d%s/result", u.Scheme, u.Host, worker.ID, task.UUID)
+			c.JSON(201, task)
+		} else {
+			c.JSON(resp.StatusCode(), respStruct)
+		}
 	} else {
-		c.JSON(resp.StatusCode(), resp.String())
+		c.JSON(resp.StatusCode(), gin.H{"error": resp.String(), "StatusCode": resp.StatusCode()})
 	}
-	// log.Info(resp.Size())
 	if action.Jobs.Cache > 0 && resp.StatusCode() == 200 && c.Request.Method == "GET" {
-		go model.SetResult(storage, resp.Body(), key, action.Jobs.Cache)
+		go models.SetResult(storage, resp.Body(), key, action.Jobs.Cache)
 	}
-	// break
 
 }