+++ middleware/routeParser.go
@@ -22,7 +22,7 @@ import (
 	"strconv"
 	s "strings"
 
-	"github.com/ezbastion/ezb_srv/model"
+	"github.com/ezbastion/ezb_srv/models"
 
 	"github.com/gin-gonic/gin"
 	log "github.com/sirupsen/logrus"
@@ -31,11 +31,13 @@ import (
 // use cache for sql https://github.com/goenning/go-cache-demo/blob/master/cache/cache.go
 func RouteParser(c *gin.Context) {
 	tr, _ := c.Get("trace")
-	trace := tr.(model.EzbLogs)
+	trace := tr.(models.EzbLogs)
 	logg := log.WithFields(log.Fields{
 		"middleware": "RouteParser",
 		"xtrack":     trace.Xtrack,
 	})
+	logg.Debug("start")
+
 	escapedPath := c.Request.URL.EscapedPath()
 	path := s.Split(escapedPath, "/")
 	// fmt.Printf("escapedPath: '%d'\n", len(path))
@@ -80,18 +82,18 @@ func RouteParser(c *gin.Context) {
 		action := path[3]
 		access := s.ToUpper(c.Request.Method)
 		viewApis, _ := c.Get("ViewApi")
-		ViewApis := viewApis.([]model.ViewApi)
+		ViewApis := viewApis.([]models.ViewApi)
 		apiFound := false
-		var CurrentAction []model.ViewApi
+		var CurrentAction []models.ViewApi
 		for i := range ViewApis {
 			if ViewApis[i].Ctrlver == version {
-				// fmt.Println(version)
+				logg.Debug("ViewApis version match:", version)
 				if ViewApis[i].Ctrl == ctrl {
-					// fmt.Println(ctrl)
+					logg.Debug("ViewApis ctrl match:", ctrl)
 					if ViewApis[i].Action == action {
-						// fmt.Println(action)
+						logg.Debug("ViewApis action match:", action)
 						if s.ToUpper(ViewApis[i].Access) == access {
-							// fmt.Println(c.Request.Method)
+							logg.Debug("ViewApis access match:", access)
 							apiFound = true
 							// CurrentAction = ViewApis[i]
 							CurrentAction = append(CurrentAction, ViewApis[i])
@@ -110,7 +112,7 @@ func RouteParser(c *gin.Context) {
 		// fmt.Println(CurrentAction)
 		// c.Set("CurrentAction", CurrentAction)
 		apiPath, _ := c.Get("apiPath")
-		ApiPath := apiPath.([]model.ApiPath)
+		ApiPath := apiPath.([]models.ApiPath)
 		mapping := make(map[int]string)
 		for _, a := range ApiPath {
 			// fmt.Println(a)