+++ CHANGELOG.md
@@ -2,10 +2,22 @@
 
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
+## 0.2.1 - 2019-10-22
+- set tokenid to anonymous when no Authorization header found
+- Fix empty output, return 204 "no content"
+- use struc interface between srv & wks
+ 
+## 0.2.0 - 2019-10-12
+- Add polling support for long tasks
 
+## 0.1.3 - 2019-06-19
+- Fix models namespace
+- Fix logrotate file loc
+- Add icon
+- Add windows file info
 
 ## 0.1.2 - 2019-05-26
-Move to AGPLv33 opensource.
+Move to AGPLv3 opensource.
 
 ## 0.1.1 - 2019-01-07
 Fisrt public release